--- a/.gitignore
+++ b/.gitignore
@@ -11,6 +11,9 @@
 build/
 dist/
 
+# for autocomplete
+compile_commands.json
+
 # Pytest verbose output
 test-results/