--- a/xformers/csrc/attention/cuda/fmha/kernel_backward.h
+++ b/xformers/csrc/attention/cuda/fmha/kernel_backward.h
@@ -591,7 +591,7 @@ struct AttentionBackwardKernel {
         // multiple preloads, dropout Zij tile, and 3 stages push us over shared
         // memory capacity on A100. set a ceiling on number of stages to save
         // shared memory if dropout is in use.
-        kPreloadMmas && kApplyDropout
+        kPreloadMmas && kApplyDropout && (kBlockSizeI * kBlockSizeJ > 64 * 64)
             ? cutlass::const_min(2, DefaultConfig::kStages)
             : DefaultConfig::kStages, // Stages
         false, // SplitKSerial