--- a/xformers/csrc/attention/cuda/fmha/kernels/backward_f32_aligned_dropout.cu
+++ b/xformers/csrc/attention/cuda/fmha/kernels/backward_f32_aligned_dropout.cu
@@ -1,0 +1,8 @@
+// This file is auto-generated. See "generate_kernels.sh"
+#ifndef XFORMERS_MEM_EFF_ATTENTION_DISABLE_BACKWARD
+#include "../kernel_backward.h"
+INSTANTIATE_ATTENTION_KERNEL_BACKWARD_SM50(float, true, true);
+INSTANTIATE_ATTENTION_KERNEL_BACKWARD_SM70(float, true, true);
+INSTANTIATE_ATTENTION_KERNEL_BACKWARD_SM75(float, true, true);
+INSTANTIATE_ATTENTION_KERNEL_BACKWARD_SM80(float, true, true);
+#endif