… CHANGELOG.md
… tests/test_mem_eff_attention.py
… xformers/benchmarks/benchmark_mem_eff_attention.py
… xformers/csrc/attention/cuda/fmha/kernel_forward.h
… xformers/csrc/attention/cuda/fmha/kernels/backward_bf16_dropout_k64.cu
… xformers/csrc/attention/cuda/fmha/kernels/backward_bf16_k128.cu
… xformers/csrc/attention/cuda/fmha/kernels/backward_bf16_k64.cu
… xformers/csrc/attention/cuda/fmha/kernels/backward_f16.cu
… xformers/csrc/attention/cuda/fmha/kernels/backward_f16_aligned.cu
… xformers/csrc/attention/cuda/fmha/kernels/backward_f16_aligned_dropout.cu
… xformers/csrc/attention/cuda/fmha/kernels/backward_f16_aligned_dropout_k128.cu
… xformers/csrc/attention/cuda/fmha/kernels/backward_f16_aligned_dropout_k64.cu
… xformers/csrc/attention/cuda/fmha/kernels/backward_f16_aligned_k128.cu
… xformers/csrc/attention/cuda/fmha/kernels/backward_f16_aligned_k64.cu
… xformers/csrc/attention/cuda/fmha/kernels/backward_f16_dropout.cu
… xformers/csrc/attention/cuda/fmha/kernels/backward_f16_dropout_k128.cu
… xformers/csrc/attention/cuda/fmha/kernels/backward_f16_dropout_k64.cu
… xformers/csrc/attention/cuda/fmha/kernels/backward_f16_k128.cu
… xformers/csrc/attention/cuda/fmha/kernels/backward_f16_k64.cu
… xformers/csrc/attention/cuda/fmha/kernels/backward_f32.cu
… xformers/csrc/attention/cuda/fmha/kernels/backward_f32_aligned.cu
… xformers/csrc/attention/cuda/fmha/kernels/backward_f32_aligned_dropout_k128.cu
… xformers/csrc/attention/cuda/fmha/kernels/backward_f32_aligned_dropout_k64.cu
… xformers/csrc/attention/cuda/fmha/kernels/backward_f32_aligned_k128.cu
… xformers/csrc/attention/cuda/fmha/kernels/backward_f32_aligned_k64.cu
… xformers/csrc/attention/cuda/fmha/kernels/backward_f32_dropout.cu
… xformers/csrc/attention/cuda/fmha/kernels/backward_f32_dropout_k128.cu
… xformers/csrc/attention/cuda/fmha/kernels/backward_f32_dropout_k64.cu
… xformers/csrc/attention/cuda/fmha/kernels/backward_f32_k128.cu
… xformers/csrc/attention/cuda/fmha/kernels/backward_f32_k64.cu