## What does this PR do?
Adds support for attention bias, bias gradient, and dropout to CUTLASS FlashAttention.

i'm mostly new to CUDA programming and totally new to CUTLASS so please let me know if i'm doing anything that doesn't make sense, is slow, or is otherwise weird :)

one note: i noticed the CPU and pure CUDA implementations also support attention bias, but they [expect bias to be same across queries](https://github.com/facebookresearch/xformers/blob/main/xformers/components/attention/csrc/cuda/mem_eff_attention/attention.cu#L736). Bias here is implemented to accept different values along rows

### TODOs
- general optimization - in particular storing bias gradients is quite slow (benchmarks below). i've done some light profiling/optimization so far but hopefully room for more improvements
- correctness + performance testing on SM80 - i only have a T4 (SM75) to work with, will try to find an A100 to do some testing with
- [probably another PR] bias types like ALIBI probably could be implemented more efficiently without a tensor bias (compute on the fly instead of having to load a large matrix from global memory). starting with this as a more general solution

there's a bug in the backward benchmarks that prevent me from using it to test performance (doesn't seem to be caused by my PR as same thing happens in main). will see if i can find a fix.

```python
Traceback (most recent call last):
  File "/home/ubuntu/ws/xformers/xformers/benchmarks/benchmark_mem_eff_attention.py", line 343, in <module>
    benchmark_main_helper(benchmark_backward, CASES, min_run_time=min_run_time)
  File "/home/ubuntu/ws/xformers/xformers/benchmarks/utils.py", line 436, in benchmark_main_helper
    measurement = benchmark_object.blocked_autorange(
  File "/home/ubuntu/.conda/envs/xformers/lib/python3.10/site-packages/torch/utils/benchmark/utils/timer.py", line 394, in blocked_autorange
    number = self._estimate_block_size(min_run_time)
  File "/home/ubuntu/.conda/envs/xformers/lib/python3.10/site-packages/torch/utils/benchmark/utils/timer.py", line 311, in _estimate_block_size
    time_taken = self._timeit(number)
  File "/home/ubuntu/.conda/envs/xformers/lib/python3.10/site-packages/torch/utils/benchmark/utils/timer.py", line 256, in _timeit
    return max(self._timer.timeit(number), 1e-9)
  File "/home/ubuntu/.conda/envs/xformers/lib/python3.10/timeit.py", line 178, in timeit
    timing = self.inner(it, self.timer)
  File "<timeit-src>", line 6, in inner
  File "/home/ubuntu/.conda/envs/xformers/lib/python3.10/site-packages/torch/_tensor.py", line 396, in backward
    torch.autograd.backward(self, gradient, retain_graph, create_graph, inputs=inputs)
  File "/home/ubuntu/.conda/envs/xformers/lib/python3.10/site-packages/torch/autograd/__init__.py", line 173, in backward
    Variable._execution_engine.run_backward(  # Calls into the C++ engine to run the backward pass
  File "/home/ubuntu/.conda/envs/xformers/lib/python3.10/site-packages/torch/autograd/function.py", line 253, in apply
    return user_fn(self, *args)
  File "/home/ubuntu/.conda/envs/xformers/lib/python3.10/site-packages/torch/autograd/function.py", line 399, in wrapper
    outputs = fn(ctx, *args)
  File "/home/ubuntu/ws/xformers/xformers/ops/fmha/__init__.py", line 96, in backward
    query, key, value, out, lse, rng_state, attn_bias_tensor = ctx.saved_tensors
RuntimeError: one of the variables needed for gradient computation has been modified by an inplace operation: [torch.cuda.FloatTensor [384, 1, 224]] is at version 1; expected version 0 instead. Hint: enable anomaly detection to find the operation that failed to compute its gradient, with torch.autograd.set_detect_anomaly(True).
```


but here's some results from the benchmarks i wrote for myself. the labels take the form
```
(batch_sz, seq_len, n_heads, head_dim, dtype, bias_shape, bias_requires_grad, dropout_p)
```
```
[-------------------------------------------- attn --------------------------------------------]
                                                                        |  reference  |  cutlass
1 threads: -------------------------------------------------------------------------------------
      (8, 512, 64, 128, torch.float16, (512, 512, 512), False, 0.0)     |     12.4    |     7.4
      (8, 512, 64, 128, torch.float16, (512, 512, 512), False, 0.5)     |     15.4    |     9.0
      (8, 512, 64, 128, torch.float16, (512, 512, 512), True, 0.0)      |     12.5    |     7.5
      (8, 512, 64, 128, torch.float16, (512, 512, 512), True, 0.5)      |     15.5    |     9.1
      (8, 512, 64, 128, torch.float16, None, False, 0.0)                |     10.1    |     5.9
      (8, 512, 64, 128, torch.float16, None, False, 0.5)                |     12.6    |     7.4
      (8, 1024, 64, 128, torch.float16, (512, 1024, 1024), False, 0.0)  |     44.0    |    28.8
      (8, 1024, 64, 128, torch.float16, (512, 1024, 1024), False, 0.5)  |     55.0    |    34.7
      (8, 1024, 64, 128, torch.float16, (512, 1024, 1024), True, 0.0)   |     44.6    |    28.9
      (8, 1024, 64, 128, torch.float16, (512, 1024, 1024), True, 0.5)   |     55.0    |    34.8
      (8, 1024, 64, 128, torch.float16, None, False, 0.0)               |     36.5    |    22.2
      (8, 1024, 64, 128, torch.float16, None, False, 0.5)               |     45.2    |    28.5

Times are in milliseconds (ms).

[------------------------------------------ attn-bwd ------------------------------------------]
                                                                        |  reference  |  cutlass
1 threads: -------------------------------------------------------------------------------------
      (8, 512, 64, 128, torch.float16, (512, 512, 512), False, 0.0)     |     19.2    |    24.0
      (8, 512, 64, 128, torch.float16, (512, 512, 512), False, 0.5)     |     19.2    |    24.5
      (8, 512, 64, 128, torch.float16, (512, 512, 512), True, 0.0)      |     22.2    |    28.7
      (8, 512, 64, 128, torch.float16, (512, 512, 512), True, 0.5)      |     22.2    |    29.0
      (8, 512, 64, 128, torch.float16, None, False, 0.0)                |     19.3    |    22.7
      (8, 512, 64, 128, torch.float16, None, False, 0.5)                |     19.3    |    23.3
      (8, 1024, 64, 128, torch.float16, (512, 1024, 1024), False, 0.0)  |     62.6    |    90.9
      (8, 1024, 64, 128, torch.float16, (512, 1024, 1024), False, 0.5)  |     62.6    |    93.3
      (8, 1024, 64, 128, torch.float16, (512, 1024, 1024), True, 0.0)   |     74.5    |   109.5
      (8, 1024, 64, 128, torch.float16, (512, 1024, 1024), True, 0.5)   |     74.8    |   110.5
      (8, 1024, 64, 128, torch.float16, None, False, 0.0)               |     62.7    |    85.2
      (8, 1024, 64, 128, torch.float16, None, False, 0.5)               |     62.7    |    87.5
```

## Before submitting

- [Y] Did you have fun?
  - Make sure you had fun coding 🙃
- [Y] Did you read the [contributor guideline](https://github.com/facebookresearch/xformers/blob/master/CONTRIBUTING.md)?
- [N] Was this discussed/approved via a Github issue? (no need for typos, doc improvements)
- [N/A] Did you make sure to update the docs?
- [Y] Did you write any new necessary tests?
- [Y] Did you update the [changelog](https://github.com/facebookresearch/xformers/blob/master/CHANGELOG.md)? (if needed)


## PR review
Anyone in the community is free to review the PR once the tests have passed.
If we didn't discuss your PR in Github issues there's a high chance it will not be merged.