@danthe3rd 2022-12-14
Hi @jfc4050 

First, that's a pretty impressive PR and significant changes, which must have taken a lot of effort :o
Having attention bias (and in a smaller measure dropout) is something we've wanted to implement for some time but didn't have time to work on, so this contribution is very welcome!
I'll try to take a look at this in the next few days - as it's a pretty big chunk.

The main things that will matter (in order of importance) for me would be:
(1) Is there any performance regression for the users who don't use a mask/dropout (mostly worried for V100/A100 - I'll run benchmarks there)
(2) The forward part has been [upstreamed to CUTLASS](https://github.com/NVIDIA/cutlass/tree/master/examples/41_fused_multi_head_attention), and @hwu36 is working on optimising it. It would be great to avoid too much divergence between both files, we need to figure out something.
(3) Ease of use (eg compile time + binary size): as this multiplies by 2 the total number of attention kernels with the additional Dropout specialisation
(4) Compatibility with older hardware (eg not just Sm75+)

Once again, thanks a lot for putting all of this work there!