@jfc4050 2022-12-14
> you did the effort to do things properly

thats a relief to hear :sweat_smile: